enhance(console): Clean up implementation of Command and GeneratorCommand enhance(console): Seperate clearLine, exec, task, and title into traits chore(console): Remove unused Schedule class
